requirements.txt
urllib3==2.2.3
python-dotenv==1.0.1
python-dateutil==2.9.0
praw==7.8.1
toml==0.10.2